reorganized codebase into parts
