+++ cmd/api/main.go
@@ -1,0 +1,20 @@
+package main
+
+import (
+	"log"
+	"net/http"
+	"os"
+    "github.com/SimplQ/simplQ-golang/internal/mux"
+)
+
+func main() {
+	listenAddr := ":8080"
+	if val, ok := os.LookupEnv("FUNCTIONS_CUSTOMHANDLER_PORT"); ok {
+		listenAddr = ":" + val
+	}
+
+    mux.InitalizeRoutes()
+
+	log.Printf("About to listen on %s. Go to https://127.0.0.1%s/", listenAddr, listenAddr)
+	log.Fatal(http.ListenAndServe(listenAddr, nil))
+}
