+++ internal/handler/hellohandler.go
@@ -1,0 +1,15 @@
+package handler
+
+import (
+    "fmt"
+    "net/http"
+)
+
+func HelloHandler(w http.ResponseWriter, r *http.Request) {
+	message := "This HTTP triggered function executed successfully. Pass a name in the query string for a personalized response.\n"
+	name := r.URL.Query().Get("name")
+	if name != "" {
+		message = fmt.Sprintf("Hello, %s. This HTTP triggered function executed successfully.\n", name)
+	}
+	fmt.Fprint(w, message)
+}
