+++ internal/mux/mux.go
@@ -1,0 +1,10 @@
+package mux
+
+import (
+    "net/http"
+    "github.com/SimplQ/simplQ-golang/internal/handler"
+)
+
+func InitalizeRoutes() {
+	http.HandleFunc("/api/HttpExample", handler.HelloHandler);
+}
